+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## [next]
 
 * fix: Pop the menu flyout before than calling the close callback ([#1009](https://github.com/bdlukaa/fluent_ui/issues/1009))
-* fix: `Button.child` is now automatically centered ([#1011](https://github.com/bdlukaa/fluent_ui/issues/1011))
+* fix: `Button.child` is now vertically centered ([#1011](https://github.com/bdlukaa/fluent_ui/issues/1011))
 * fix: Adjust `Button`'s foreground color when hovered and pressed ([#1012](https://github.com/bdlukaa/fluent_ui/issues/1012))
+* feat: Added `.textController` and `.onTextChanged` to `EditableComboBox`. ([#1017](https://github.com/bdlukaa/fluent_ui/issues/1017))
 
 ## 4.8.5